+++ CHANGELOG.rst
@@ -1,3 +1,4 @@
+
 [X.X.X]
 --------
 
@@ -5,6 +6,7 @@ Added:
 ^^^^^^
 * Added option to disable hard filter of variants in matched normal https://github.com/Clinical-Genomics/BALSAMIC/pull/1509
 * Added check to verify sample sex for all workflows https://github.com/Clinical-Genomics/BALSAMIC/pull/1516
+* SOR filter to WGS TN SNV quality filter https://github.com/Clinical-Genomics/BALSAMIC/pull/1506
 * GT field to IGH-DUX4 variant https://github.com/Clinical-Genomics/BALSAMIC/pull/1527
 * ONC field annotations from Clinvar https://github.com/Clinical-Genomics/BALSAMIC/pull/1527
 
@@ -29,7 +31,6 @@ Fixed:
 ^^^^^^
 * Merged VarDict and TNscope variants now correctly show both callers in FOUND_IN info field https://github.com/Clinical-Genomics/BALSAMIC/pull/149
 * Fixed somalier container https://github.com/Clinical-Genomics/BALSAMIC/pull/1538
-
 
 [16.0.1]
 -------